+++ source/_asnx_lib_/sensors/dht.h
@@ -23,6 +23,9 @@
 
 
 /***** DEFINES ********************************************************/
+/* Enable last-measurement time check (requires systick lib) */
+#define DHT_CHECK_LAST_MEAS     (1)
+
 /* Timing-related */
 #define DHT_TIMING_MIN_INTERVAL 2000UL
 #define DHT_TIMING_TIMEOUT      200